feat(app): enhance notification parsing, data consistency, and UI polish
Major improvements:
- Redesign NotificationListener with better error handling and retry logic (+243 lines)
- Enhance PaymentParser with improved regex patterns and blacklist filtering (+159 lines)
- Add data validation and consistency checks across Pro/Standard flavors
- Polish UI components (settings, trash screen, monthly chart)
- Update network security configuration for better compatibility
- Add comprehensive README with download badges and version comparison table
- Include custom icon quick-start guide (Chinese documentation)
- Add development plans, design specs, and pyproject.toml for tooling

Stats: 27 files changed, +1577 insertions, -204 deletions

+++ app/src/main/java/com/example/autobookkeeper/ui/components/MonthlyBarChart.kt
@@ -236,7 +236,7 @@ fun MonthlyBarChart(
             }
 
             val entries = data.mapIndexed { index, item ->
-                BarEntry(index.toFloat(), item.amount, item)
+                BarEntry(index.toFloat(), item.amount.toFloat(), item)
             }
 
             val dataSet = BarDataSet(entries, "").apply {
